posts/views/viewset.py
from posts.models import Post
from posts.serializers import PostListSerializer
from rest_framework import viewsets


class PostViewSet(viewsets.ModelViewSet):
    queryset = Post.objects.all()
    serializer_class = PostListSerializer

