requirements.txt
asgiref==3.2.7
Django==3.0.5
django-extensions==2.2.9
django-redis==4.11.0
djangorestframework==3.11.0
pytz==2020.1
redis==3.5.0
six==1.14.0
sqlparse==0.3.1
